Parse a float from the compass

+++ sailbot/arduino.py
@@ -35,7 +35,7 @@ class Arduino(object):
 
     def get_compass(self):
         """Get the heading from the compass"""
-        return int(self.__sendCommand('c'))
+        return float(self.__sendCommand('c'))
 
     def get_thermometer(self):
         """Get the temperature in degrees Celsius"""
